separate usage summary and details + delete unused component
